Replace "Result" type with "Boolean".
Boolean is like Bool, but can carry meta-data for each branch of the
sum type. It is more generic than Result, which had a fixed type of
meta-data for the "No" (~ False) branch which was slightly "HTTP-ish".

Also lift the constraint, where we could only work with Snap rather
than being polymorphic on MonadSnap.