Add :||: in addition to :|:.
So far :|: was returning the meta-data as Either (TVal a) (TVal b) to
allow mixed-types. This commit re-defines :|: to require equal types for
both branches. The motivation is that this will be the primary use case
and doesn't force clients to pattern match on Eithers.

The new :||: connective is the old :|:, i.e. allows different types in
both cases and bundles them in an Either, aka :+:.