Introduce optional delta value to Boolean.
Sometimes it is not enough to evaluate predicates per route and pick the
first one which evaluates to 'T'. For example content negotiation
requires picking the handler which is closest to the requested media
type. In order to support these scenarios, the 'T' branch of 'Boolean'
contains a list of (ByteString, Double) pairs. The string is a
classifier to denote some arbitrary category the delta applies to, the
Double represents the distance or delta from some optimum, e.g. 0.

During predicate evaluation the ones without any delta will be preferred
as they have no distance to the optimum. Otherwise the closest, i.e. the
one whose sum of delta values is lowest will be selected. Only the
intersection of all categories will be considered.

The previous paragraph explains this commit's implementation which is
not perfect. Whereas previously  predicates would be evaluated one by
one until one matches, the order is no longer guaranteed. Also a single
'Const ()' will always win, i.e. less specific predicates are favoured
which is not what we want.