Rename Readable to Convert and change parsing.
Readable supports parsing of bytestrings which contain language level
literals, e.g. a Char value is surrounded by single quotes, a Text
value by double quotes. While this supports relatively straightforward
parsing of recursive data structures such as lists, it is often not what
is needed in the context of request headers or parameters. Here
converting from ByteString to for instance Text means little more than
decoding the bytes. It is not expected that the parameter values are
wrapped in quotes. Thus I revert the implementation of Readable to its
former restricted variant with minor improvements.