Replace media-types with type literals.
I.e. instead of writing `MediaType Application Json` one would write
`Media "application" "json"`. Also this enables the simplification of
the `Accept` and `ContentType` predicate insofar as it is sufficient to
mention the predicate name. The type parameters will be infered from the
handler signature:

    sitemap :: Routes Snap ()
    sitemap = do
        post  "/image" image1 ContentType
        post  "/image" image2 ContentType

    image1 :: Media "image" "png" -> Snap ()
    image1 = writeBS . pack . ("image1: " ++) . show

    image2 :: Media "image" "gif" -> Snap ()
    image2 = writeBS . pack . ("image2: " ++ ) . show

This is still experimental and needs further exploration.